backend/requirements.txt
# FastAPI e servidor
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Banco de dados
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.1

# Validação e configuração
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Autenticação
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Pagamentos
stripe>=7.0.0

# Testes
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0

# Utilitários
python-dateutil>=2.8.0

